Render strings as well
Encounter a memory limit bug. Basically the memory would become garbage
and return an ordering of undefined, causing nothing to be rendered.
Reducing the draw array size fixed this issue.

I believe we were reaching the memory limit of the wasm module and it
was reallocating where the arrays were?